Conflicts:
	Applications/CLI/tubeAnisotropicDiffusiveDeformableRegistration/Testing/CMakeLists.txt
	Applications/DICOMSeriesConverter/Testing/CMakeLists.txt
	Applications/ImageMath/Testing/CMakeLists.txt
	MIDAS_Keys/Boxes_normals.mhd.md5
	MIDAS_Keys/Boxes_weights.mhd.md5
	MIDAS_Keys/Sphere_normals.mhd.md5
	MIDAS_Keys/Sphere_normalsMultiresolution.mhd.md5
	MIDAS_Keys/Sphere_weights.mhd.md5
	MIDAS_Keys/Sphere_weightsMultiresolution.mhd.md5
	MIDAS_Keys/SyntheticSurfaceImage.mha.md5
	MIDAS_Keys/SyntheticTubeImage.mha.md5
	MIDAS_Keys/SyntheticTubePlusSurfaceImage.mha.md5
	MIDAS_Keys/SyntheticTubePlusSurfaceImage_oriented.mha.md5
	MIDAS_Keys/im0001.vk.mha.md5

+++ Base/Numerics/tubeSplineApproximation1D.cxx
@@ -29,7 +29,7 @@ SplineApproximation1D::
 SplineApproximation1D()
 : Spline1D()
 {
-  m_SplineApproximation1DMatrixConst = (float)(1.0/6.0);
+  m_SplineApproximation1DMatrixConst = (double)(1.0/6.0);
   m_SplineApproximation1DMatrix(0, 0) = 1;
   m_SplineApproximation1DMatrix(0, 1) = 0;
   m_SplineApproximation1DMatrix(0, 2) = 0;
@@ -54,7 +54,7 @@ SplineApproximation1D( UserFunc<int, double> *newFunval,
   Optimizer1D *newOpt1D )
 : Spline1D(newFunval, newOpt1D)
 {
-  m_SplineApproximation1DMatrixConst = (float)(1.0/6.0);
+  m_SplineApproximation1DMatrixConst = (double)(1.0/6.0);
   m_SplineApproximation1DMatrix(0, 0) = 1;
   m_SplineApproximation1DMatrix(0, 1) = 0;
   m_SplineApproximation1DMatrix(0, 2) = 0;
